Refactor files names to include Activity
